feat: add payment functionality to purchase and sales invoices
- Implemented AddPayment component for handling payments in both PurchaseList and SalesList.
- Updated usePurchaseList and useSalesList hooks to manage selected invoice and payment modal state.
- Enhanced PurchaseList and SalesList components to include payment buttons for unpaid invoices.
- Modified useAddPurchase and useAddSales hooks to track invoice status and due amounts.
- Refactored AddSales component to include payment status and funds selection for paid invoices.
- Created AddPayment modal for processing payments with fund selection and notes.

+++ src/backend/db.js
@@ -49,8 +49,8 @@ CREATE TABLE IF NOT EXISTS customers (
   name TEXT,
   phone TEXT,
   address TEXT,
-  total REAL DEFAULT 0,
-  total_paid REAL DEFAULT 0,
+  total INTEGER DEFAULT 0,
+  total_paid INTEGER DEFAULT 0,
   createdAt TEXT DEFAULT (datetime('now'))
 )
 `,
@@ -63,8 +63,8 @@ CREATE TABLE IF NOT EXISTS suppliers (
   name TEXT,
   phone TEXT,
   address TEXT,
-  total REAL DEFAULT 0,
-  total_paid REAL DEFAULT 0,
+  total INTEGER DEFAULT 0,
+  total_paid INTEGER DEFAULT 0,
   createdAt TEXT DEFAULT (datetime('now'))
 )
 `,
